// packages/workflow/src/schematics/advanced-crud/schema.ts
/**
 * @license
 * SPDX-License-Identifier: MIT
 *
 * Copyright (c) 2020-2026 Houssemeddine LAOUITI (Wiforge)
 * https://www.wiforge.com
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root of this repository.
 */

export interface AdvancedCRUDOption {
  // The name of the target entity.
  ename: string

  // The name of the related entity service : Default = {$name}Service (ex. UserService).
  sname: string

  // The path to create the entity.
  path?: string

  // The name of the project.
  project?: string

  // the project root path
  projectRootPath: string
}